Fix setting of database password
First, do not set the database password in the crowbar service model if
it's not set, but leave the cookbook set it like we do in other
cookbooks (and like the cookbook was trying to do).

Second, set the correct password early, in the common recipe, before we
define the sql connection variable, instead of later on (in the database
recipe).

This fixes a crash on first run of the cookbook, as seen in
http://lists.us.dell.com/pipermail/crowbar/2013-June/003354.html