Make quantum services work on very first start
We had a lot of subscribes/notifies with no delay (:immediately) to
start/restart quantum services as soon as possible. The goal of this was
to have quantum run fine before we hit the post_install_conf recipe.

However, the first start/restart were broken because all configuration
files are not created at this point.

Instead, we create some hack where we create a list of services that
have to be restarted, and we restart them before entering the
post_install_conf recipe. This will work fine because the relevant
actions in that recipe are run in the Execute phase of the chef run, and
we restart the services in the Execute phase too (so after we collected
what will need to be restarted in the Compile phase), but before
entering the recipe.