Move cisco switch port assignment ui into proposal view
As the information managed in the topology view it closely tied to a single
quantum proposal it make good sense to also handle it as a part of the
proposal.
This also means that the nexus switch port assignments are now stored
in the quantum proposal instead of (ab)using the node objects for that.
This makes creating the config files much easier.