Enable and start the RabbitMQ service
Fixes a regression introduced by:
https://github.com/crowbar/barclamp-rabbitmq/pull/14

On SUSE, other than on Ubuntu the rabbitmq service is not automatically enabled
and started after the package installation. So this needs to be done by the
cookbook. There shouldn't be the need to special case this for SUSE as the
:enable and :start action shouldn't do anything if the service is already
enabled and running (as on Ubuntu).

Additionally this sets the correct path to the rabbitmq-plugins binary on SUSE
platforms.