Re-enable "vlan" mode for openvswitch
Force "use_vlan" to false on the nova_fixed network when using vlan modes.

When ovs is used in vlan-mode quantums sets up an ovs-bridge to do the vlan
tagging. That bridge needs be attached to a none VLAN interface in order to
work correctly.

When linux-bridge is used quantum's linux bridge agent will create the correct
vlan interfaces automatically.

For the openvswitch vlan mode there is the limitation, that the bridge for the
openvswitch network needs to be on top a different physical device (or bond)
than the "admin" network.  Otherwise nodes might not come up correctly after a
reboot. Currently the quantum UI will display a warning about that.

A working openvswitch vlan mode is a requirement for the cisco nexus switch
support.