Switch to the ml2 plugin for osv and linuxbridge based setups
The "old" ovs and linuxbridge plugins will be removed in Icehouse so it's
better to switch now. Featurewise this doesn't add anything new. E.g. it's
currently not possible to select multiple type_drivers for ml2.

This commit currently breaks cisco support. (Will be fixed again by a later
commit)