Update passwords migration script to remove node attributes
On update, we want to use the db password (and metering secret) from the
proposal / role, and not the ones that might have been generated on the node
(through the cookbook). As the passwords are a "normal" attribute on the node,
but a "default" attribute on the role, the old ones from the node is used.
Which means we really need to remove them to get the new ones used.

To minimize config changes, we actually import the passwords from one
node (which should be good enough in most cases as we allow only one
ceilometer proposal, so only one node should have the passwords).