Split the HA code in two recipes for more reliability
Because postgresql has config files not in /etc but in
/var/lib/pgsql/data (which is inside the mount that we use for
postgresql storage when in HA mode), the setup is quite tricky to get
right.

We must first setup the storage bits of the HA, so that we can write the
configuration files in the right place.

Once we have the configuration files inside the mount, then we can
configure (and therefore start) the postgresql pacemaker resource.
Otherwise, we're having a race, with postgresql starting before the
mount is completed on this node.

In addition to that, we used to rely on postgresql starting to generate
some of the files (thanks to initdb). We now have to do that manually.

In the non HA mode, we didn't have that many issues with the order,
because we do not use a mount which might not be there when creating the
configuration files.