Run once neutron-db-manage check_migration to do the db_sync
Neutron is actually magic and can perform that by itself, except that in
the HA case, concurrent neutron instances try to do it at the same time,
and fail. That's why we need to manually do it first.

Also, we only do it the very first time (and we remember that through an
attribute).