ha: Fix initial DB sync
The initial database sync for neutron doesn't occur with
"neutron-db-manage check_migration"; it occurs when the neutron-server
daemon starts. This is perfect for non-HA, but breaks for HA since
cluster members will run this concurrently and will fail to execute the
db migrations all at the same time.

To make this work, I can't think of any other way than starting once
neutron-server (to get the db sync) and then stopping it, to then hand
over all of this to pacemaker. We only do this for HA (since it's
useless without HA) and we only do this the first time on the cluster
founder (like before, for the old code we thought would work).