ha: Fix race on creation of drbd resources
The drbd OCF resource agent expects the drbd primitive and the drbd ms
resource to be created in the same transaction. When it doesn't, then
the drbd primitve will fail to be monitored, which can lead to fencing.

We didn't see that all the time because we create the ms resource fast,
so most of the time, it just works.