Switch to thin LVM backend by default
This provides first of all significant more performance
for volumes about snapshots (thick snapshots turn every write
on the volume to a sync+flush for the snapshot), and saves disk
space at the same time. Furthermore it avoids hang issues
with the SLE11 lvm (which does not have lvmetad enabled and hence
rescans the lvs on every command).