Fix races caused by initial DB population by neutron-server with HA
The DB migration that we do with neutron-db-manage is not doing
everything: on first startup, neutron-server is also populating the DB
with actual content.

This is totally racy with HA, and we end up with errors like:

  DBDuplicateEntry: (IntegrityError) duplicate key value violates unique constraint "ml2_gre_allocations_pkey"
or
  (IntegrityError) duplicate key value violates unique constraint "pg_type_typname_nsp_index"

To avoid this, in the HA case, we simply go back to what we were doing
before 932d3e8 and 23aad0d: we start and stop neutron-server on the
founder node.

https://bugzilla.novell.com/show_bug.cgi?id=889325