Add time_to_live parameter into the barclamp
Add the parameter time_to_live (ceilometer.conf) into the barclamp,
so the user can decide when the meter lines are removed from the
database.

This patch was tested with a mongodb configuration.

This patch was made by Jiri (jsuchome). I only tested it and added
the parameter in the ceilometer.conf.erb template.