Do not fail when OSD nodes are temporarily off
When OSD nodes are, for instance, rebooting, they lose their ceph-osd
role. So our query to check whether ceph was deployed is kind of wrong.

If we don't find any OSD node, then we check if the ceph setup was
likely done (by checking for /etc/ceph/ceph.client.admin.keyring), and
if yes, we just print a warning.