Add support for VXLAN when using OVS (bsc#923206)
The default VNI range is 4096:99999. Rationale is that we start after
the vlan range (to avoid accidental confusion), and we create a big
range (without going to the maximum, because it shouldn't be needed in
most cases and is more readable by default).

VXLAN support with linuxbridge should be ready on the cookbook side too,
but it's not implemented in the UI yet, and it was agreed to delay that
as it only works with SLE12 while OVS works with both SLE11 and SLE12.

Note that the multicast group feature for broadcast emulation only works
with linuxbridge.

https://bugzilla.suse.com/show_bug.cgi?id=923206