Add support for DVR
DVR stands for Distributed Virtual Routers, see
https://github.com/openstack/neutron-specs/blob/master/specs/juno/neutron-ovs-dvr.rst
https://wiki.openstack.org/wiki/Neutron/DVR

The implementation is loosely based on what's in
https://wiki.openstack.org/wiki/Neutron/DVR/HowTo

For DVR, the L3 agent needs to run not just on the network node, but
also on compute nodes. So the configuration for this has to move to the
common code path.

When using HA, we still setup the L3 agent with pacemaker on the network
nodes, but it's unclear yet if this just works fine.