Add a use_l2pop attribute to enable/disable L2 population
While we want to always enable it, it turns out that doesn't work well
on upgrade where networking can break until chef-client is run on all
nodes running a L2 agent. So this can lead to a 15 minutes outage.

Therefore we have an attribute to enable this setting set to true by
default for new deployments, but that gets set to false on upgrades.