Update heat_domain_admin user when it already exists
We were only trying to create the user, and doing nothing when it
already existed. Now, we instead  update the user if it already exists.

It's a bit painful to do that because the openstack CLI is not as
helpful as it could and we need to make sure we're looking at the user
in the right domain (user listing doesn't seem to correctly filter by
domain). For that reason, we also prefer to refer to the user by its ID
instead of its name.