ceph: Only run "ceph -s" for external cluster
On initial run of recipe with crowbar cluster, ceph.conf is missing when
we try to run this. There's no way around it, so instead just assume
that the ceph cluster will work fine (so move "ceph -s" to external
cluster only). If it's not, then we have other issues as part of crowbar
anyway.