Create external networks as "flat" provider networks (bsc#946882)
This should work regardless of the mechanism driver that is used  (openvswitch
or linuxbridge) as we have the ml2 type driver "flat" enabled unconditionally.
This is also in line with the behaviour of devstack.

For linuxbridge setups this commit also adjust the interface mappings for
external networks to use the VLAN interface instead for the underlying physical
interface. That's required for "flat" networks.