Help Docker instances get proper MTU value
We already send the right MTU value to instances via an option in DHCP,
but this mechanism may not work in all Docker instances because some of
them won't even run a DHCP client. Having the right MTU setting on
instance creation is needed. There's an option (network_device_mtu)
that can be configured in nova.conf to achieve that.

The value is set to 1400 bytes to be in sync with the values provided
to the instances via DHCP options.

This is only needed when the compute node is running Docker instances
and the networking requires tunneling.