cinder: Ugly hack to make cinder quotas work again
We set [keymgr] encryption_auth_url in cinder.conf to the keystone v3
auth URI. This is terrible because that's unrelated to this setting and
this hardcodes a need for the v3 API, but that's really upstream bugs.

See upstream bugs:
  https://bugs.launchpad.net/cinder/+bug/1517043
  https://bugs.launchpad.net/cinder/+bug/1516085
  https://bugs.launchpad.net/cinder/+bug/1511960
  https://bugs.launchpad.net/cinder/+bug/1509308

On top of this, the quotas are broken when using insecure SSL with
keystone, but it's not something we can fix here.