Fix interface mappings when floating is not on a VLAN (bsc#955811)
If "nova_floating" as "use_vlan" turned of and is configured to use the same
conduit as "nova_fixed" they need to be on the same "physnet" in neutron. Try
to find those cases, write the interface_mapping accordingly and adjust the
"net-create" parameters for that.

Partiall fix for: https://bugzilla.suse.com/show_bug.cgi?id=955811