rabbitmq: Fix nodename used for rabbitmq
It was missing the rabbit@ part, which is wrong as this could mean a
conflict with other erlang apps.

Note that this change breaks existing deployment of rabbitmq, as
rabbitmq can't get restarted (due to the nodename change). So it's only
okay to get in if we know that the upgrade process will stop rabbitmq.