rabbitmq: Constrain services to only run on controllers when doing HA
As seen in some logs, when there are two resources (A and B) with a
colocation constraint, and only one controller-only location constraint
for (for resource B), it seems that pacemaker sometimes still tries to
start resource A on a non-controller member of the cluster. This is
clearly wrong, so we need to add additional controller-only location
constraints in that case.

That means that for the first transaction (the one related to the
storage), we need one location constraint for the fs primitive and, if
using drbd, one location constraint for the ms.

For the second transaction (the one related to the service), if we use
drbd, then we also need a location constraint for the vip and for the
service. But if we don't use drbd, then we have a group and one location
constraint is actually enough in that case.