nova: Change grub config after xen kernel is installed
The code to change the grub config is parsing the list of available
kernels, but since it was running in the compilation phase, it had to
reboot twice to actually work as intended (first time because this
command is run before the xen kernel is installed, so we trigger a
reboot without rebooting in the xen kernel).

Instead, run this in the convergence phase with a ruby_block.