Removes public endpoint from keystone conf (bsc#954957)
How things work on the side of keystoneclient:
1. It queries OS_AUTH_URL (keystone) for version discovery.
OS_AUTH_URL (keystone) returns a response with an url, which is
public_endpoint if it is set or host_url if it is not set.
There is a comment about that in keystone code:

url = context['host_url']

2. The client passes this url to an auth plugin.

3. The auth plugin appends something (/auth/token) to the url
and queries the new url

Removing the public endpoint fixes the issue of mismatched endpoints.