nova: Install libvirt, nova & neutron packages on corosync nodes
When doing HA for compute nodes, the corosync nodes still need to have
everything needed to setup the resource agents that will run on compute
nodes. That means that init scripts, service files, etc. that will be
used to define primitives must be installed on corosync nodes.

Also, fence_compute requires /var/log/nova; while we could create the
directory ourselves, we would also need to create a nova user and group
(to be sure there's no permission problem for the directory in case nova
is installed later). We might as well just install the package that does
this for us.