nova: Fix sync marks for compute HA
Corosync nodes should not be required to be in the nova proposal and
therefore might not have the revision of the nova proposal directly in
their attributes. So the default way to fetch the revision for the sync
marks might not work.

Therefore, we fetch the revision used for the sync marks from the compute
nodes.