nova: Let crowbar orchestrate compute nodes HA
This is really pushing the limits of crowbar orchestration, and
therefore requires a small hack.

In a way similar to how we change element_order in the pacemaker
barclamp depending on whether there are remote nodes or not, for the
nova barclamp we change the elements and element_order in case we have
remote nodes in clusters acting as compute nodes.

Our requirement is that after the setup has been done on the compute
nodes that are remote nodes (installing and configuring nova packages),
we need to run the compute_ha recipe on the corosync nodes. To achieve
this, we do two things when applying:

 - we add a new set of elements for the nova-compute-ha role: these are
   the clusters that are also part of compute roles; this helps make
   sure
 - we change element_order to include as last batch the nova-compute-ha
   role

The result is that after the compute roles have been applied, we will
apply the compute-ha role.

The benefit of doing this just before apply_role (in active_update) is
that we benefit from all the magic in apply_role: checking if the nodes
are ready, marking them as applying, getting a lock on chef-client, etc.
If we were doing any other hack to implement this, all this logic would
not apply to the corosync nodes and this could create various issues.