nova: Change the "eq controller" rule for location constrain to "ne compute"
When we add a cluster to a compute role, the remote nodes become compute
nodes. However, there is no guarantee that the corosync nodes are
assigned to any controller role of openstack. So the corosync nodes
might not have "controller" as OpenStack-role.

A safe approach is simply to enforce the resources to not run on nodes
that have "compute" as OpenStack-role.