nova: Rename nova-compute-ha role to nova-ha-compute
We have several places in the code that use the following regular
expressions to find nodes with compute roles: /^nova-compute-/

Unfortunately, that also matches nova-compute-ha. While we could fix all
the code doing the regxp to discard nova-compute-ha, new code with a
similar regexp could be introduced without the discarding code.

So better be safe and rename the role to avoid this.