Start Agents before Neutron-Server
There is known issue with Neutron right now that when we allocate
a new network, it tries to bind the port for the Neutron router
right away. However, if at that point the agent isn't running yet,
the port will go into failed state and not be retried and the
neutron network is broken then.

So lets try to start the agents before neutron-server so that
we can bind the ports on the host.