manila: Fix default proposal for generic backend
The default proposal has a generic backend with no password nor key;
however, the UI makes you believe that a password is set (it actually
displays "undefined").

To have this work, we need to change the validation to not check that
there's one attribute between password and key, but also to check that
there's one non-empty attribute.