neutron, nova: Use CrowbarPacemakerService for compute nodes HA services
We cannot just have the services started by the standard service
provider, as the services are handled by pacemaker. So we use
CrowbarPacemakerService.

We also use the new restart_crm_resource flag for nova-compute: this
flag is required for services matching a pacemaker resource which is
defined with an OCF agent: in that case, using standard service
primitives wouldn't work.