Run flavor migration tool when upgrading from juno
When upgrading from juno to kilo, after running the db migrations the flavor
information for instance had be migrated using the "nova-manage db
migrate_flavor_data" tool. As we're upgrading directly from juno to liberty we
need execute this step else. Because of this the actually db migration step is
now split into three steps:
1. Run db migrations up to kilo
2. run flavor migration tool
3. Run the rest of the db migrations

To have this working a patched nova is needed (migrate_flavor_data forward
ported from kilo to juno).