neutron: Fix broken MTU for tunnel networks after upgrade (bsc#968426)
When upgrading from Juno, we move the MTU advertisement from
/etc/neutron/dnsmasq-neutron.conf (custom dnsmasq config file) to new
settings in configuration files.

However, the MTU for each network is saved in the database and the
database migration from Juno to Liberty just set this to NULL, resulting
in broken MTU.

This is only an issue for networks with a tunnel type. In tex/Juno, we
only had two such possibilities: gre and vxlan (and with ovs only).

We fix this by tweaking the MTU in the database directly and restarting
the DHCP agents, which results in restarting dnsmasq for each network
with the correct settings.

https://bugzilla.suse.com/show_bug.cgi?id=968426