Use a random z/VM instance root password by default
We change the default for zvm_config_drive_inject_password
to true which causes the root password to be set to a randomly generated
individual password rather than the hardcoded root-pass. In case
this is undesired this parameter is made configurable so you can change
it back to a globally hardcoded password (that is now also configurable)
instead.

(cherry picked from commit 6b4111c5c35f219e47c031f7d4df4354d7807ce9)