nova: Allow custom vendordata for metadata server
Create a vendordata.json file which will be served through the metadata
server to the guest VM.
This vedordata.json contains can i.e. contains some information about a
possible SMT server. The provided information could then used by the
SUSE guest images to automatically configure the repositories for the
guest.
This is based on a solution the public cloud team is already using.