first attempt at setting up mention-bot
This is just a test setup for https://github.com/facebook/mention-bot
but it's hopefully close to what we would want as a final configuration,
although obviously we would probably want to add many more entries for
all the OpenStack components and the team members working on them.

See https://github.com/facebook/mention-bot#configuration for more
detail on the meaning of the configuration options.

Extended glob syntax is used; see:

  https://github.com/facebook/mention-bot/issues/140