trove: Make trove able to use radosgw instead of swift
This also means we move away from get_instance for fetching the swift
proxy, as get_instance always falls back to the current node if it
doesn't find a result. So we need a real search.

Also note that we do not have any attribute specifying which ceph
environment to use, hence the call to search instead of
search_env_filtered.