only mention SUSE employees for now
as we don't have any contributors outside the SUSE org it does not
make sense to mention people that haven't contributed for a long
time.
This prevents spamming for now, and if crowbar grows to form an
infrastructure for a wider community we can adapt the list again.
Main reason for this change is to stop spamming non-active
contributors.