chef: Use Barclamp::Inventory.Network objects
We do this instead of accessing attributes, in order to centralize the
location of the network definitions, so it can be changed.

We have to be careful for neutron related networks, since these are not
necessarily enabled on the nodes, so we need to access the definition in
the role directly.