trove: Replace openstack-identity with keystone_register
This is a first step to unify the trove cookbook with the rest of the
cookbooks.
Remove the openstack-identity cookbook and replace it with the keystone_register
methods to have a single resource provider handling the user/service/endpoint
registration. This also removes the usage of the admin-token.
Note: The reason why the code to register the endpoints and users moved to api.rb
is that this is an ongoing effort to also get rid of openstack-database and
have the chef recipies all in the trove cookbook. And all the other services are
doing the endpoint/service/user handling in the api.rb recipe.