cinder: Allow to use a HA cluster for cinder-volume in active/passive
We cannot do active/active at the moment because there needs to be some
synchonization between the various volume services in that case.

But it turns out that active/passive is relatively easy to achieve: it's
only a matter of making sure that the cinder-volume service on the
various cluster members register with the same host. We use the
vhostname of the cluster for this.

Of course, it doesn't work with the LVM backend, so we have validation
checks for this in Crowbar.

Due to the fact that we need to change the host in the cinder config,
and because it would impact cinder-controller too, we do the config
change in a cinder-volume specific config file. Otherwise, we would
impact the cinder-scheduler service, and  only one cinder-scheduler
service would be register (with the vhostname of the cluster), which is
wrong in this case.

(cherry picked from commit 0c87c97f8ed9dec479102ceb6e965c7208a5ae4f)