nova: Allow defining on which network live migration happens
Right now, live migration is being done only on the admin network.
However, the admin network might be on a 1G interface, while some other
network might be defined with 10G interface. It'd be better to use such
a network for live migration, to speed up things.

We add a new attribute to specify which network to use.

It's important to keep in mind that the public network is not
appropriate here, as it would make the compute nodes publicly reachable,
and libvirt listen on the public IP. So it's better to define a new
private network, or use an existing one like the storage network.