glance: Remove enable_v3_api switch
There will be no glance V3 API. Instead there is another service called
glance-glare (Glance Artifacts Repository) which has it's own service
endpoint.
From the glance docs:
"""
The Artifacts API can’t be the Images v3 API since Artifacts are not the
same as Images. Conceptually, a virtual machine image could be an
Artifact, and the Glare code has been designed to be compatible with the
Images v2 API. But at this time, there are no plans to implement an
Images v3 API.
"""