trove: Add public IP allocation
Allocate a public IP in the Trove barclamp so it will also work if it is
deployed on a standalone machine without other barclamps. Without this code it
won't work if it is deployed to a machine that doesn't have at least one other
barclamp assigning a public IP.

Note: since this commit adds a apply_role_pre_chef_call() function to the Trove
barclamp's model it also contains commented HA code to make enabling HA for
Trove easier later on.