Heat: Change clients endpoints on publicURL
We would like to change endpoints on publicURL for Heat
and Magnum clients. This will be used by OS::Heat::WaitCondition
resource and wc_notify.

(cherry picked from commit 74611f7e49911eba0e13b8e020d5af3f12194229)

As part of the cherry-pick, we only take the Heat part (as Magnum
support is not in this branch).