swift: remove swift-dispersion-populate timeout
swift-dispersion-populate can be slow, making some HA sync marks
timeout. The current timeout of the syncs are 60 seconds, and for
this command is 120 seconds.

This commit change the way we call the command, delegating to an
unattached process.  This solution is an intermediate step before
converting it into a service.

Fix bsc#989683