nova: Allow to define a custom network for xCAT access (bsc#976778)
xCAT needs to be able to ssh to the compute nodes. Allow to use
something different then the admin network by setting the
nova.zvm.zvm_xcat_network to the name of the network that is defined in
the network barclamp.

z/VM will then allocate an IP in that network and set "my_ip" in
nova.conf accordingly.

https://bugzilla.suse.com/show_bug.cgi?id=976778