neutron: Simplify code to get nova_fixed interface
We're using the Barclamp::Inventory.Network class now.

Note: Barclamp::Inventory.Network.interface() returns the interface name
including the vlan tag if use_vlan is true for the network while
Barclamp::Inventory.Network.lookup_interface_info didn't include it.
This is not a real problem for "nova_fixed". As "use_vlan" is always set to
"false" when the "nova_fixed" network is enabled on any node (see
NeutronService.rb).