magnum: Use correct port for endpoint
This commit sets the port used for Magnum's service catalog entries to
node[:magnum][:bind_port], as opposed to the port retrieved from MagnumHelper
that was used before this commit. The port retrieved from MagnumHelper is the
service's actual listening port, which is incorrect in an HA setting. There,
the endpoint catalog entries must use the port where the haproxy instance
forwarding to the actual service port listens.