ceilometer: Use auth_type to password (bsc#1001273)
Ceilometer 7.0 change the section service_credentials.  If auth_type
is not set, some queries to keytone will use the legacy v2.0 API,
and this cause problems (POST can be blocked in SSL connections).

If we set this value to `password` V3 will be used.

Update `username` parameter to `user-name`, to avoid deprecation
warning.