cinder: Add configuration option for multipath (bsc#997607)
Currently, mulitpath support is set by default in the cinder and
nova cookbooks, and both cookbooks force it to be true, but the
built package overrides nova's setting to false.

Remove that default support, and add a new option to cinder which allows
operators to enable or disable multipath support for both cinder
and nova.