don't break whole cluster when not upgrading
https://github.com/crowbar/crowbar-openstack/pull/562 totally broke HA
cluster setup, because when the pre-upgrade node attribute is not
present, the location constraints generated by this helper prevent
anything useful from running.

So instead we reverse the

    pre-upgrade eq false

logic to

    pre-upgrade ne true

so that the constraints will *only* prevent resources from running
when pre-upgrade is set to true, and not when it isn't set at all.