aodh: Switch to WSGI app behind mod_wsgi
The builtin wsgi server in Newton does not read the port setting
from the config file anymore, so any attempt at setting that (which
is needed to be customized for HA) is failing. Put it behind our
standard apache2/mod_wsgi setup that should allow this to be configured
in apache2 now instead.