neutron: Don't search for nova attributes when running on compute node
Some neutron recipes are included from the nova cookbook, to setup
neutron bits on compute nodes.

This is signaled through a hack with the node[:cookbook] attribute
(which is set to "nova"), and in that case, we know that the node is a
node part of the nova barclamp and that has, therefore, access to the
attributes of the nova proposal. So we don't need to search for a nova
controller to get these attributes.