ha: Re-order rabbitmq UID/GID and pacemaker resource creation
We reset the UID/GID of rabbitmq to `91` to make it consistent across
the cluster. This change has to happen before we create the fs-rabbitmq
pacemaker resource, otherwise rabbitmq can not be started due to
potentially wrong ownership values on its home directory. This patch
moves the `bash` resource further up.