keystone: Fix endpoint race generator
The cached client side urls were depending on keystone server role
on the server finishing first to set the attributes so that they could
be read. This is not always the case, especially with HA the non-HA
nodes are a lot faster than the HA nodes. So locally generate the
proper urls and cache the correct ones.